# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KG Extraction Backend",
    "description": "接收前端上传的 PDF，提取文本并保存为 TXT 文件",
    "version": "1.0.0"
  },
  "paths": {
...
    "/graph/extract": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Extract Graph",
        "operationId": "extract_graph_graph_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/extract": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Api Extract Graph",
        "operationId": "api_extract_graph_api_graph_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph/clean": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Clean Graph",
        "operationId": "clean_graph_graph_clean_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphCleanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/clean": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Api Clean Graph",
        "operationId": "api_clean_graph_api_graph_clean_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphCleanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph/merge": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Merge Graph",
        "operationId": "merge_graph_graph_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/merge": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Api Merge Graph",
        "operationId": "api_merge_graph_api_graph_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph/relations/clean": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Clean Graph Relations",
        "operationId": "clean_graph_relations_graph_relations_clean_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphRelationCleanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/relations/clean": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Api Clean Graph Relations",
        "operationId": "api_clean_graph_relations_api_graph_relations_clean_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphRelationCleanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph/quality/report": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Graph Quality Report",
        "operationId": "graph_quality_report_graph_quality_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphQualityReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/quality/report": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Api Graph Quality Report",
        "operationId": "api_graph_quality_report_api_graph_quality_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphQualityReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph/final/generate": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Final Graph Generate",
        "operationId": "final_graph_generate_graph_final_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalGraphGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/graph/final/generate": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Api Final Graph Generate",
        "operationId": "api_final_graph_generate_api_graph_final_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalGraphGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/pipeline/pdf-to-graph": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Pdf To Final Graph",
        "operationId": "pdf_to_final_graph_pipeline_pdf_to_graph_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_pdf_to_final_graph_pipeline_pdf_to_graph_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/pdf-to-graph": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Api Pdf To Final Graph",
        "operationId": "api_pdf_to_final_graph_api_pipeline_pdf_to_graph_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_pdf_to_final_graph_api_pipeline_pdf_to_graph_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_api_pdf_to_final_graph_api_pipeline_pdf_to_graph_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "auto"
          },
          "entity_types": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Types"
          },
          "relations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relations"
          },
          "allowed_entity_types": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Entity Types"
          },
          "allowed_relation_types": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Relation Types"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_api_pdf_to_final_graph_api_pipeline_pdf_to_graph_post"
      },
...
      "Body_pdf_to_final_graph_pipeline_pdf_to_graph_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "auto"
          },
          "entity_types": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Types"
          },
          "relations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relations"
          },
          "allowed_entity_types": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Entity Types"
          },
          "allowed_relation_types": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Relation Types"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_pdf_to_final_graph_pipeline_pdf_to_graph_post"
      },
...
      "FinalGraphGenerateRequest": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          }
        },
        "type": "object",
        "required": [
          "extraction_id"
        ],
        "title": "FinalGraphGenerateRequest"
      },
      "GraphCleanRequest": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          }
        },
        "type": "object",
        "required": [
          "extraction_id"
        ],
        "title": "GraphCleanRequest"
      },
      "GraphExtractionRequest": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "auto"
          },
          "allowed_entity_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Entity Types"
          },
          "allowed_relation_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Relation Types"
          }
        },
        "type": "object",
        "required": [
          "extraction_id"
        ],
        "title": "GraphExtractionRequest"
      },
      "GraphMergeRequest": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          }
        },
        "type": "object",
        "required": [
          "extraction_id"
        ],
        "title": "GraphMergeRequest"
      },
      "GraphQualityReportRequest": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          }
        },
        "type": "object",
        "required": [
          "extraction_id"
        ],
        "title": "GraphQualityReportRequest"
      },
      "GraphRelationCleanRequest": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          }
        },
        "type": "object",
        "required": [
          "extraction_id"
        ],
        "title": "GraphRelationCleanRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}